pyproject.toml
[project]
name = "tritonadmix"
version = "0.1.0"
description = "A pure Python CLI tool for admixture inference using EM algorithm"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "numpy>=1.26.0",
    "scikit-allel>=1.3.0",
]

[project.scripts]
tritonadmix = "tritonadmix.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv]
package = true
